Draw point by map position, fixed bug with block position

+++ include/Tetris/maprenderer.h
@@ -26,9 +26,9 @@ namespace Tetris {
             std::shared_ptr<Gfx::Engine> engine;
             std::shared_ptr<Map> map;
             SDL_Rect area;
-            int tileSize;
+            float tileSize;
 
-            int calculateTileSize();
+            float calculateTileSize();
             void drawBlock(TileColors color, int x, int y);
             void drawBlock(TileColors color, SDL_Point p);
     };
